Update readme to include instructions to load rest-client
Since https://github.com/jayniz/rspec-api-test/blob/master/lib/rspec_api_test.rb doesn't require `rest-client` by default, the execution fails if not manually loaded as a dependency. This is the easiest solution, just to help other people coming to this gem, but of course the proper one would be requiring the gem in the entry .rb script.